requirements.txt
s3fs
scikit-learn
pandas < 3.0.0
hyperopt
mlflow

xgboost
catboost
seaborn
matplotlib

ipykernel

fastapi
uvicorn[standard]
python-dotenv